refactor: split code into smaller components for better readability
